ProbeSet-class {affy}	R Documentation

Class ProbeSet

Description

A simple class that contains the PM and MM data for a probe set from one or more samples.

Objects from the Class

Objects can be created by applying the method probeset to instances of AffyBatch.

Slots

id:

Object of class "character" containing the probeset ID.

pm:

Object of class "matrix" containing the PM intensities. Columns represent samples and rows the different probes.

mm:

Object of class "matrix" containing the MM intensities.

Methods

colnames

signature(x = "ProbeSet"): the column names of the pm matrices which are the sample names

express.summary.stat

signature(x = "ProbeSet", pmcorrect = "character", summary = "character"): applies a summary statistic to the probe set.

sampleNames

signature(object = "ProbeSet"): the column names of the pm matrices which are the sample names.

Note

More details are contained in the vignette.

See Also

probeset, AffyBatch-class

Examples

if (require(affydata)) {
  data(Dilution)
  ps <- probeset(Dilution, geneNames(Dilution)[1:2])
  names(ps)
  print(ps[[1]])
}

[Package affy version 1.62.0 Index]