affy-deprecated {affy}	R Documentation

Deprecated functions in package ‘affy’

Description

These functions are provided for compatibility with older versions of affy only, and will be defunct at the next release.

Details

The following functions are deprecated and will be made defunct; use the replacement indicated below:

loess.normalize: normalize.loess

maffy.normalize

multiloess

simplemultiLoess

[Package affy version 1.62.0 Index]