affy-options {affy}	R Documentation

Options for the affy package

Description

Description of the options for the affy package.

Note

The affy package options are contained in the Bioconductor options. The options are:

use.widgets: a logical used to decide on the default of widget use.

compress.cel: a logical

compress.cdf: a logical

probes.loc: a list. Each element of the list is it self a list with two elements what and where. When looking for the informations about the locations of the probes on the array, the elements in the list will be looked at one after the other. The first one for which what and where lead to the matching locations information is used. The element what can be one of package, environment or file. The element where depends on the corresponding element what.

if package: location for the package (like it would be for the argument lib.loc for the function library.)

if environment: an environment to look for the information (like the argument env for the function get).

if file: a character with the path in which a CDF file can be found.

Examples

## get the options
opt <- getOption("BioC")
affy.opt <- opt$affy

## list their names
names(affy.opt)

## set the option compress.cel
affy.opt$compress.cel <- TRUE
options(BioC=opt)

[Package affy version 1.62.0 Index]