bg.adjust {affy}	R Documentation

Background adjustment (internal function)

Description

An internal function to be used by bg.correct.rma.

Usage

bg.adjust(pm, n.pts = 2^14, ...)
bg.parameters(pm, n.pts = 2^14)

Arguments

pm
a pm matrix

n.pts
number of points to use in call to density.

...
extra arguments to pass to bg.adjust.

Details

Assumes PMs are a convolution of normal and exponential. So we observe X+Y where X is background and Y is signal. bg.adjust returns E[Y|X+Y, Y>0] as our background corrected PM. bg.parameters provides ad hoc estimates of the parameters of the normal and exponential distributions.

Value

a matrix

See Also

bg.correct.rma

[Package affy version 1.64.0 Index]