bg.correct {affy}	R Documentation

Background Correction

Description

Background corrects probe intensities in an object of class AffyBatch.

Usage

bg.correct(object, method, ...)

bg.correct.rma(object,...)
bg.correct.mas(object, griddim)
bg.correct.none(object, ...)

Arguments

object
An object of class AffyBatch.

method
A character that defines what background correction method will be used. Available methods are given by bg.correct.methods.

griddim
grid dimension used for mas background estimate. The array is divided into griddim equal parts. Default is 16.

...
arguments to pass along to the engine function.

Details

The name of the method to apply must be double-quoted. Methods provided with the package are currently:

bg.correct.none: returns object unchanged.

bg.correct.chipwide: noise correction as described in a ‘white paper’ from Affymetrix.

bg.correct.rma: the model based correction used by the RMA expression measure.

They are listed in the variable bg.correct.methods. The user must supply the word after "bg.correct", i.e none, subtractmm, rma, etc...

More details are available in the vignette.

R implementations similar in function to the internal implementation used by bg.correct.rma are in bg.adjust.

Value

An AffyBatch for which the intensities have been background adjusted. For some methods (RMA), only PMs are corrected and the MMs remain the same.

Examples

if (require(affydata)) {
  data(Dilution)

  ##bgc will be the bg corrected version of Dilution
  bgc <- bg.correct(Dilution, method="rma") 

  ##This plot shows the tranformation
  plot(pm(Dilution)[,1],pm(bgc)[,1],log="xy",
  main="PMs before and after background correction")
}

[Package affy version 1.62.0 Index]