cdfFromBioC {affy}	R Documentation

Functions to obtain CDF files

Description

A set of functions to obtain CDF files from various locations.

Usage

cdfFromBioC(cdfname, lib = .libPaths()[1], verbose = TRUE)
cdfFromLibPath(cdfname, lib = NULL, verbose=TRUE)
cdfFromEnvironment(cdfname, where, verbose=TRUE)

Arguments

cdfname
name of the CDF.

lib
install directory for the CDF package.

where
environment to search.

verbose
logical controlling extra output.

Details

These functions all take a requested CDF environment name and will attempt to locate that environment in the appropriate location (a package's data directory, as a CDF package in the .libPaths(), from a loaded environment or on the Bioconductor website. If the environment can not be found, it will return a list of the methods tried that failed.

Value

The CDF environment or a list detailing the failed locations.

Author(s)

Jeff Gentry

[Package affy version 1.62.0 Index]