cleancdfname {affy}	R Documentation

Clean Affymetrix's CDF name

Description

This function converts Affymetrix's names for CDF files to the names used in the annotation package and in all Bioconductor.

Usage

cleancdfname(cdfname, addcdf = TRUE)

Arguments

cdfname
A character denoting Affymetrix'x CDF file name

addcdf
A logical. If TRUE it adds the string "cdf" at the end of the cleaned CDF name. This is used to name the cdfenvs packages.

Details

This function takes a CDF filename obtained from an Affymetrix file (from a CEL file for example) and convert it to a convention of ours: all small caps and only alphanumeric characters. The details of the rule can be seen in the code. We observed exceptions that made us create a set of special cases for mapping CEL to CDF. The object mapCdfName holds information about these cases. It is a data.frame of three elements: the first is the name as found in the CDF file, the second the name in the CEL file and the third the name in Bioconductor. mapCdfName can be loaded using data(mapCdfName).

Value

A character

Examples

cdf.tags <- c("HG_U95Av2", "HG-133A")
for (i in cdf.tags)
  cat(i, "becomes", cleancdfname(i), "\n")

[Package affy version 1.64.0 Index]