generateExprSet-method {affy}	R Documentation

generate a set of expression values

Description

Generate a set of expression values from the probe pair information. The set of expression is returned as an ExpressionSet object.

Usage

  computeExprSet(x, pmcorrect.method, summary.method, ...)

  generateExprSet.methods()

  upDate.generateExprSet.methods(x)

Arguments

x
a AffyBatch holding the probe level informations to generate the expression values, for computeExprSet, and for upDate.generateExprSet.methods it is a character vector..

pmcorrect.method
the method used to correct PM values (see section 'details').

summary.method
the method used to generate the expression value (see section 'details').

...
any of the options of the normalization you would like to modify.

Details

An extra argument ids= can be passed. It must be a vector of affids. The expression values will only be computed and returned for these affyids.

The different methods available through this mechanism can be accessed by calling the method generateExprSet.methods with an object of call Cel.container as an argument.

In the Affymetrix design, MM probes were included to measure the noise (or background signal). The original algorithm for background correction was to subtract the MM signal to the PM signal. The methods currently included in the package are "bg.correct.subtractmm", "bg.correct.pmonly" and "bg.correct.adjust".

To alter the available methods for generating ExprSets use upDate.generateExprSet.methods.

See Also

method generateExprSet of the class AffyBatch
expresso

Examples

if (require(affydata)) {
  data(Dilution)

  ids <- c( "1000_at","1001_at")

  eset <- computeExprSet(Dilution, pmcorrect.method="pmonly",
                         summary.method="avgdiff", ids=ids)
}

[Package affy version 1.62.0 Index]