generateExprVal.method.playerout {affy}	R Documentation

Generate an expression value from the probes informations

Description

Generate an expression from the probes

Usage

  generateExprVal.method.playerout(probes, weights=FALSE, optim.method="L-BFGS-B")

Arguments

probes
a list of probes slots from PPSet.container

weights
Should the resulting weights be returned ?

optim.method
see parameter 'optim' for the function optim

Details

A non-parametric method to weight each perfect match probe in the set and to compute a weighted mean of the perfect match values. One will notice this method only makes use of the perfect matches. (see function playerout.costfunction for the cost function).

Value

A vector of expression values.

Author(s)

Laurent <laurent@cbs.dtu.dk>
(Thanks to E. Lazaridris for the original playerout code and the discussions about it)

References

Emmanuel N. Lazaridis, Dominic Sinibaldi, Gregory Bloom, Shrikant Mane and Richard Jove A simple method to improve probe set estimates from oligonucleotide arrays, Mathematical Biosciences, Volume 176, Issue 1, March 2002, Pages 53-58

[Package affy version 1.64.0 Index]