justRMA {affy}	R Documentation

Read CEL files into an ExpressionSet

Description

Read CEL files and compute an expression measure without using an AffyBatch.

Usage

just.rma(..., filenames = character(0),
               phenoData = new("AnnotatedDataFrame"),
               description = NULL,
               notes = "",
               compress = getOption("BioC")$affy$compress.cel,
               rm.mask = FALSE, rm.outliers = FALSE, rm.extra = FALSE,
               verbose=FALSE, background=TRUE, normalize=TRUE,
               bgversion=2, destructive=FALSE, cdfname = NULL)

justRMA(..., filenames=character(0),
              widget=getOption("BioC")$affy$use.widgets,
              compress=getOption("BioC")$affy$compress.cel,
              celfile.path=getwd(),
              sampleNames=NULL,
              phenoData=NULL,
              description=NULL,
              notes="",
              rm.mask=FALSE, rm.outliers=FALSE, rm.extra=FALSE,
              hdf5=FALSE, hdf5FilePath=NULL,verbose=FALSE,
              normalize=TRUE, background=TRUE, 
              bgversion=2, destructive=FALSE, cdfname = NULL)

Arguments

...
file names separated by comma.

filenames
file names in a character vector.

phenoData
an AnnotatedDataFrame object.

description
a MIAME object.

notes
notes.

compress
are the CEL files compressed?

rm.mask
should the spots marked as 'MASKS' set to NA?

rm.outliers
should the spots marked as 'OUTLIERS' set to NA?

rm.extra
if TRUE, then overrides what is in rm.mask and rm.oultiers.

hdf5
use of hdf5 ? (not available yet)

hdf5FilePath
a filename to use with hdf5 (not available yet).

verbose
verbosity flag.

widget
a logical specifying if widgets should be used.

celfile.path
a character denoting the path ReadAffy should look for cel files.

sampleNames
a character vector of sample names to be used in the AffyBatch.

normalize
logical value. If TRUE, then normalize data using quantile normalization.

background
logical value. If TRUE, then background correct using RMA background correction.

bgversion
integer value indicating which RMA background to use 1: use background similar to pure R rma background given in affy version 1.0 - 1.0.2

2: use background similar to pure R rma background given in affy version 1.1 and above

destructive
logical value. If TRUE, then works on the PM matrix in place as much as possible, good for large datasets.

cdfname
Used to specify the name of an alternative cdf package. If set to NULL, then the usual cdf package based on Affymetrix' mappings will be used.

Details

justRMA is a wrapper for just.rma that permits the user to read in phenoData, MIAME information, and CEL files using widgets. One can also define files where to read phenoData and MIAME information.

If the function is called with no arguments justRMA(), then all the CEL files in the working directory are read, converted to an expression measure using RMA and put into an ExpressionSet. However, the arguments give the user great flexibility.

phenoData is read using read.AnnotatedDataFrame. If a character is given, it tries to read the file with that name to obtain the AnnotatedDataFrame object as described in read.AnnotatedDataFrame. If left NULL and widget=FALSE (widget=TRUE is not currently supported), then a default object is created. It will be an object of class AnnotatedDataFrame with its pData being a data.frame with column x indexing the CEL files.

description is read using read.MIAME. If a character is given, it tries to read the file with that name to obtain a MIAME instance. If left NULL but widget=TRUE, then widgets are used. If left NULL and widget=FALSE, then an empty instance of MIAME is created.

The arguments rm.masks, rm.outliers, rm.extra are passed along to the function read.celfile.

Value

An ExpressionSet object, containing expression values identical to what one would get from running rma on an AffyBatch.

Author(s)

In the beginning: James MacDonald <jmacdon@med.umich.edu> Supporting routines, maintenance and just.rma: Ben Bolstad <bmb@bmbolstad.com>

See Also

rma, read.affybatch

[Package affy version 1.64.0 Index]