loess.normalize {affy}	R Documentation

Normalize arrays

Description

This function treats PM and MM as the raw data on each chip. It fits loess curves to MVA plots and tries to normalize the chips with respect to each other by forcing log ratios to be scattered around the same constant.

Usage


loess.normalize(mat, subset = sample(1:(dim(mat)[2]), 5000), epsilon
                 = 10^-2, maxit = 1, log.it = TRUE, verbose = TRUE,
                 span = 2/3, family.loess = "symmetric")

Arguments

mat 	 a matrix with columns containing the values of the chips to normalize.

subset 	 a subset of the data to fit a loess to.

epsilon 	 small value used for the stopping criterion.

maxit 	 maximum number of iterations.

log.it 	 logical. If TRUE it takes the log2 of mat

verbose 	 logical. If TRUE displays current pair of chip being worked on.

span 	 span to be used by loess

family.loess 	 "gaussian" or "symmetric" as in loess

Details

Experience shows that you only need 1-2 iterations to obtain useful results. This function is not written in an efficient way. In order to make it faster, loess is fit to a sample of the data which we then use to predict the curve for all the data. By setting family.loess="gaussian" the function is faster, but you risk losing information on differentially expressed genes. The function normalize.quantiles is faster.

Value

A matrix with normalized values for chips in columns.

Author(s)

Rafael A. Irizarry

See Also

normalize.quantiles, maffy.normalize, maffy.subset

[Package affy version 1.18.2 Index]