normalize {affy}	R Documentation

Normalize - generic

Description

A generic function which normalizes microarray data. Normalization is intended to remove from the intensity measures any systematic trends which arise from the microarray technology rather than from differences between the probes or between the target RNA samples hybridized to the arrays.

Usage


normalize(object, ...)

Arguments

object 	 a data object containing microarray data

... 	 any other arguments

See Also

Type showMethods("normalize") at the R prompt to see what methods are available. Help on individual methods is generally available as normalize.<class> where <class> is the class of the data object. For example, for the main class in the affy package use ?normalize.AffyBatch.

Other Bioconductor packages include some related generic functions: normalizeWithinArrays, and normalizeBetweenArrays, in the LIMMA package, and maNorm in the marrayNorm package.

[Package affy version 1.18.2 Index]