normalize.constant {affy}	R Documentation

Scale probe intensities

Description

Scale array intensities in a AffyBatch.

Usage

normalize.AffyBatch.constant(abatch, refindex=1, FUN=mean, na.rm=TRUE)
normalize.constant(x, refconstant, FUN=mean, na.rm=TRUE)

Arguments

abatch
an instance of the AffyBatch-class.

x
a vector of intensities on a chip (to normalize to the reference).

refindex
the index of the array used as a reference.

refconstant
the constant used as a reference.

FUN
a function generating a value from the intensities on an array. Typically mean or median.

na.rm
parameter passed to the function FUN.

Value

An AffyBatch with an attribute "constant" holding the value of the factor used for scaling.

Author(s)

L. Gautier <laurent@cbs.dtu.dk>

See Also

AffyBatch

[Package affy version 1.64.0 Index]