normalize.contrasts {affy}	R Documentation

Normalize intensities using the contrasts method

Description

Scale chip objects in an AffyBatch-class.

Usage

normalize.AffyBatch.contrasts(abatch,span=2/3, choose.subset=TRUE,
                              subset.size=5000, verbose=TRUE,
                              family="symmetric",
                              type=c("together","pmonly","mmonly","separate"))

Arguments

abatch
an AffyBatch-class object.

span
parameter to be passed to the function loess.

choose.subset
Boolean. Defaults to TRUE

subset.size
Integer. Number of probesets to use in each subset.

verbose
verbosity flag.

family
parameter to be passed to the function loess.

type
a string specifying how the normalization should be applied.

Value

An object of the same class as the one passed.

See Also

maffy.normalize

[Package affy version 1.64.0 Index]