plotDensity {affy}	R Documentation

Plot Densities

Description

Plots the non-parametric density estimates using values contained in the columns of a matrix.

Usage

plotDensity(mat, ylab = "density", xlab="x", type="l", col=1:6,
            na.rm = TRUE, ...)

plotDensity.AffyBatch(x, col = 1:6, log = TRUE,
                      which=c("pm","mm","both"),
                      ylab = "density",
                      xlab = NULL, ...)

Arguments

mat
a matrix containing the values to make densities in the columns.

x
an object of class AffyBatch.

log
logical value. If TRUE the log of the intensities in the AffyBatch are plotted.

which
should a histogram of the PMs, MMs, or both be made?

col
the colors to use for the different arrays.

ylab
a title for the y axis.

xlab
a title for the x axis.

type
type for the plot.

na.rm
handling of NA values.

...
graphical parameters can be given as arguments to plot.

Details

The list returned can be convenient for plotting large input matrices with different colors/line types schemes (the computation of the densities can take some time).

To match other functions in base R, this function should probably be called matdensity, as it is sharing similarities with matplot and matlines.

Value

It returns invisibly a list of two matrices ‘x’ and ‘y’.

Author(s)

Ben Bolstad and Laurent Gautier

Examples

if (require(affydata)) {
  data(Dilution)
  plotDensity(exprs(Dilution), log="x")
}

[Package affy version 1.62.0 Index]