pmcorrect {affy}	R Documentation

PM Correction

Description

Corrects the PM intensities in a ProbeSet for non-specific binding.

Usage

pmcorrect.pmonly(object)

pmcorrect.subtractmm(object)

pmcorrect.mas(object, contrast.tau=0.03, scale.tau=10, delta=2^(-20))

Arguments

object
An object of class ProbeSet.

contrast.tau
a number denoting the contrast tau parameter in the MAS 5.0 pm correction algorithm.

scale.tau
a number denoting the scale tau parameter in the MAS 5.0 pm correction algorithm.

delta
a number denoting the delta parameter in the MAS 5.0 pm correction algorithm.

Details

These are the pm correction methods perfromed by Affymetrix MAS 4.0 (subtractmm) and MAS 5.0 (mas). See the Affymetrix Manual for details. pmonly does what you think: does not change the PM values.

Value

A ProbeSet for which the pm slot contains the corrected PM values.

References

Affymetrix MAS 4.0 and 5.0 manual

Examples

if (require(affydata)) {
  data(Dilution)
  gn <- geneNames(Dilution)
  pps <- probeset(Dilution, gn[1])[[1]]

  pps.pmonly <- pmcorrect.pmonly(pps)
  pps.subtractmm <- pmcorrect.subtractmm(pps)
  pps.mas5 <- pmcorrect.mas(pps)
}

[Package affy version 1.64.0 Index]