probeMatch-methods {affy}	R Documentation

Methods for accessing perfect matches and mismatches

Description

Methods for perfect matches and mismatches probes

Methods

object = AffyBatch

All the perfect match (pm) or mismatch (mm) probes on the arrays the object represents are returned.

object = ProbeSet

The pm or mm of the object are returned.

[Package affy version 1.62.0 Index]