probeNames-methods {affy}	R Documentation

Methods for accessing the Probe Names

Description

Methods for accessing Probe Names

Methods

object = Cdf

an accessor function for the name slot.

object = probeNames

returns the probe names associated with the rownames of the intensity matrices one gets with the pm and mm methods.

[Package affy version 1.62.0 Index]