summary {affy}	R Documentation

Probe Set Summarizing Functions

Description

These were used with the function express, which is no longer part of the package. Some are still used by the generateExprVal functions, but you should avoid using them directly.

See Also

expresso

[Package affy version 1.64.0 Index]