AffyBatch-class {affy}	R Documentation

Class AffyBatch

Description

This is a class representation for Affymetrix GeneChip probe level data. The main component are the intensities from multiple arrays of the same CDF type. It extends eSet.

Objects from the Class

Objects can be created using the function read.affybatch or the wrapper ReadAffy.

Slots

cdfName:

Object of class character representing the name of CDF file associated with the arrays in the AffyBatch.

nrow:

Object of class integer representing the physical number of rows in the arrays.

ncol:

Object of class integer representing the physical number of columns in the arrays.

assayData:

Object of class AssayData containing the raw data, which will be at minimum a matrix of intensity values. This slot can also hold a matrix of standard errors if the 'sd' argument is set to TRUE in the call to ReadAffy.

phenoData:

Object of class AnnotatedDataFrame containing phenotypic data for the samples.

annotation

A character string identifying the annotation that may be used for the ExpressionSet instance.

featureData

Object of class AnnotatedDataFrame containing feature-level (e.g., probeset-level) information.

experimentData:

Object of class "MIAME" containing experiment-level information.

notes:

Object of class "character" Vector of explanatory text

Extends

Class "eSet", directly.

Methods

cdfName

signature(object = "AffyBatch"): Obtains the cdfName slot.

pm<-

signature(object = "AffyBatch"): replaces the perfect match intensities

pm

signature(object = "AffyBatch"): extracts the pm intensities.

mm<-

signature(object = "AffyBatch"): replaces the mismatch intensities.

mm

signature(object = "AffyBatch"): extracts the mm intensities.

probes

signature(object = "AffyBatch", which): extract the perfect match or mismatch probe intensities. Uses which can be "pm" and "mm".

exprs

signature(object = "AffyBatch"): extracts the expression matrix.

exprs<-

signature(object = "AffyBatch", value = "matrix"): replaces the expression matrix.

se.exprs

signature(object = "AffyBatch"): extracts the matrix of standard errors of expression values, if available.

se.exprs<-

signature(object = "AffyBatch", value = "matrix"): replaces the matrix of standard errors of expression values.

[<-

signature(x = "AffyBatch"): replaces subsets.

[

signature(x = "AffyBatch"): subsets by array.

boxplot

signature(x = "AffyBatch"): creates a boxplots of log base 2 intensities.
hist

signature(x = "AffyBatch"): creates a plot showing all the histograms of the pm,mm or both data. See plot.density

computeExprSet

signature(x = "AffyBatch", summary.method = "character"): For each probe set computes an expression value using summary.method.

featureNames

signature(object = "AffyBatch"): return the probe set names also referred to as the Affymetrix IDs. Notice that one can not assign featureNames. You must do this by changing the cdfenvs.

geneNames

signature(object="AffyBatch'"): deprecated, use featureNames

getCdfInfo

signature(object = "AffyBatch"): retrieve the environment that defines the location of probes by probe set.

image

signature(x = "AffyBatch"): creates an image for each sample.

indexProbes

signature(object = "AffyBatch", which = "character"): returns a list with locations of the probes in each probe set. The affyID corresponding to the probe set to retrieve can be specified in an optional parameter genenames. By default, all the affyIDs are retrieved. The names of the elements in the list returned are the affyIDs. which can be "pm", "mm", or "both". If "both" then perfect match locations are given followed by mismatch locations.
signature(object = "AffyBatch", which = "missing") (i.e., calling indexProbes without a "which" argument) is the same as setting "which" to "pm".

intensity<-

signature(object = "AffyBatch"): a replacement method for the exprs slot, i.e. the intensities.

intensity

signature(object = "AffyBatch"): extract the exprs slot, i.e. the intensities.

length

signature(x = "AffyBatch"): returns the number of samples.

pmindex

signature(object = "AffyBatch"): return the location of perfect matches in the intensity matrix.

mmindex

signature(object = "AffyBatch"): return the location of the mismatch intensities.

dim

signature(x = "AffyBatch"): Row and column dimensions.

ncol

signature(x = "AffyBatch"): An accessor function for ncol.

nrow

signature(x = "AffyBatch"): an accessor function for nrow
.

normalize

signature(object = "AffyBatch"): a method to normalize. The method accepts an argument method. The default methods is specified in package options (see the main vignette).

normalize.methods

signature(object = "AffyBatch"): returns the normalization methods defined for this class. See normalize.

probeNames

signature(object = "AffyBatch"): returns the probe set associated with each row of the intensity matrix.

probeset

signature(object = "AffyBatch",genenames=NULL, locations=NULL): Extracts ProbeSet objects related to the probe sets given in genenames. If an alternative set of locations defining pms and mms a list with those locations should be passed via the locations argument.

bg.correct

signature(object = "AffyBatch", method="character")
applies background correction methods defined by method.
updateObject

signature(object = "AffyBatch", ..., verbose=FALSE): update, if necessary, an object of class AffyBatch to its current class definition. verbose=TRUE provides details about the conversion process.

Note

This class is better described in the vignette.

See Also

related methods merge.AffyBatch, pairs.AffyBatch, and eSet

Examples


## load example
data(affybatch.example)

## nice print
print(affybatch.example)

pm(affybatch.example)[1:5,]
mm(affybatch.example)[1:5,]

## get indexes for the PM probes for the affyID "A28102_at" 
mypmindex <- pmindex(affybatch.example,"A28102_at")
## same operation using the primitive
mypmindex <- indexProbes(affybatch.example, which="pm", genenames="A28102_at")[[1]]
## get the probe intensities from the index
intensity(affybatch.example)[mypmindex, ]

## load bigger example (try 'help(Dilution)' )
data(affybatch.example)
description(affybatch.example) ##we can also use the methods of eSet
sampleNames(affybatch.example)
abstract(affybatch.example)

[Package affy version 1.18.2 Index]