AffyRNAdeg {affy}	R Documentation

Function to assess RNA degradation in Affymetrix GeneChip data.

Description

Uses ordered probes in probeset to detect possible RNA degradation. Plots and statistics used for evaluation.

Usage


AffyRNAdeg(abatch,log.it=TRUE)
summaryAffyRNAdeg(rna.deg.obj,signif.digits=3)
plotAffyRNAdeg(rna.deg.obj, transform = "shift.scale", cols = NULL,
                 ...)

Arguments

abatch 	 An object of class AffyBatch-class.

log.it 	 A logical argument: If log.it=T, then probe data is log2 tranformed

rna.deg.obj 	 Output from AffyRNAdeg

signif.digits 	 Number of significant digits to show.

transform 	 Possible choices are "shift.scale","shift.only", and "neither". "Shift" vertically staggers the plots for individual chips, to make the display easier to read. "Scale" normalizes so that standard deviation is equal to 1.

cols 	 A vector of colors for plot, length = number of chips

... 	 further arguments for plot function.

Details

Within each probeset, probes are numbered directionally from the 5' end to the 3' end. Probe intensities are averaged by probe number, across all genes. If log.it=FALSE and transform="Neither", then plotAffyRNAdeg simply shows these means for each chip. Shifted and scaled versions of the plot can make it easier to see.

Value

AffyRNAdeg returns a list with the following components:

sample.names 	 names of samples, derived from affy batch object

means.by.number 	 average intensity by probe position

ses 	 standard errors for probe position averages

slope 	 from linear regression of means.by.number

pvalue 	 from linear regression of means.by.number

Author(s)

Leslie Cope

Examples


data(affybatch.example)
RNAdeg<-AffyRNAdeg(affybatch.example)
plotAffyRNAdeg(RNAdeg)


[Package affy version 1.18.2 Index]