MAplot {affy}	R Documentation

Relative M vs. A plots

Description

Create boxplots of M or M vs A plots. Where M is determined relative to a specified chip or to a pseudo-median reference chip.

Usage


MAplot(object,...)
Mbox(object,...)
ma.plot(A, M, subset = sample(1:length(M), min(c(10000, length(M)))),
show.statistics=TRUE, span=2/3, family.loess="gaussian", cex = 2,plot.method=c("normal","smoothScatter","add"),add.loess = TRUE,lwd = 1, lty = 1, loess.col = "red",...)

Arguments

object 	 An AffyBatch-class

... 	 Additional parameters for the routine

A 	 A vector to plot along the horizonal axis

M 	 A vector to plot along vertical axis

subset 	 A set of indices to use when drawing the loess curve

show.statistics 	 If true some summary statistics of the M values are drawn

span 	 span to be used for loess fit.

family.loess 	 "guassian" or "symmetric" as in loess.

cex 	 Size of text when writing summary statistics on plot

plot.method 	 a string specifying how the plot is to be drawn. "normal" plots points, "smoothScatter" uses the smoothScatter function. Specifying "add" means that the MAplot should be added to the current plot

add.loess 	 add a loess line to the plot

lwd 	 width of loess line

lty 	 line type for loess line

loess.col 	 color for loess line

See Also

mva.pairs

Examples


   data(affybatch.example)
   MAplot(affybatch.example)
   Mbox(affybatch.example)

[Package affy version 1.18.2 Index]