affy.scalevalue.exprSet {affy}	R Documentation

Scale normalization for expreSets

Description

Normalizes expression values using the method described in the Affymetrix user manual.

Usage


affy.scalevalue.exprSet(eset, sc = 500, analysis="absolute")

Arguments

eset 	 An ExpressionSet object.

sc 	 Value at which all arrays will be scaled to.

analysis 	 Should we do absolute or comparison analysis, although "comparison" is still not implemented.

Details

This is function was implemented from the Affymetrix technical documentation for MAS 5.0. It can be downloaded from the website of the company. Please refer to this document for details.

Value

A normalized ExpressionSet

Author(s)

Laurent

[Package affy version 1.18.2 Index]