bg.adjust {affy}	R Documentation

Background ajustment (internal function)

Description

An internal function to be used by bg.correct.rma.

Usage


bg.adjust(pm, n.pts = 2^14, ...)
bg.parameters(pm, n.pts = 2^14)

Arguments

pm 	 a pm matrix

n.pts 	 number of points to use in call to density.

... 	 extra arguments to pass to bg.adjust.

Details

Assumes PMs are a convolution of normal and exponentional. So we observe X+Y where X is backround and Y is signal. bg.adjust returns E[Y|X+Y, Y>0] as our backround corrected PM. bg.parameters provides adhoc estimates of the parameters of the normal and exponential distributions.

Value

a matrix

See Also

bg.correct.rma

[Package affy version 1.18.2 Index]