cdfFromBioC {affy}	R Documentation

Functions to obtain CDF files

Description

A set of functions to obtain CDF files from various locations.

Usage


cdfFromBioC(cdfname, lib = .libPaths()[1], verbose = TRUE)
cdfFromData(cdfname, pkg, verbose=TRUE)
cdfFromLibPath(cdfname, lib = NULL, verbose=TRUE)
cdfFromEnvironment(cdfname, where, verbose=TRUE)

Arguments

cdfname 	 The CDF desired

lib 	 Directory to install the CDF package to

pkg 	 What package to search

where 	 What environment to search

verbose 	 Controls extra output

Details

These functions all take a requested CDF environment name and will attempt to locate that environment in the appropriate location (a package's data directory, as a CDF package in the .libPaths(), from a loaded environment or on the Bioconductor website. If the environment can not be found, it will return a list of the methods tried that failed.

Value

The CDF environment or a list detailing the failed locations.

Author(s)

Jeff Gentry

[Package affy version 1.18.2 Index]