expresso {affy}	R Documentation

From raw probe intensities to expression values

Description

Goes from raw probe intensities to expression values

Usage


expresso(
        afbatch,
    # background correction
        bg.correct = TRUE,
        bgcorrect.method = NULL,
        bgcorrect.param = list(),
    # normalize
        normalize = TRUE,
        normalize.method = NULL,
        normalize.param = list(),
    # pm correction
        pmcorrect.method = NULL,
        pmcorrect.param = list(),
    # expression values
        summary.method = NULL,
        summary.param = list(),
        summary.subset = NULL,
    # misc.
        verbose = TRUE,
        widget = FALSE)

Arguments

afbatch 	 An AffyBatch object

bg.correct 	 a boolean to express whether background correction is wanted or not.

bgcorrect.method 	 the name of the background adjustment method

bgcorrect.param 	 a list of parameters for bgcorrect.method (if needed/wanted)

normalize 	 normalization step wished or not.

normalize.method 	 the normalization method to use

normalize.param 	 a list of parameters to be passed to the normalization method (if wanted).

pmcorrect.method 	 the name of the PM adjustement method

pmcorrect.param 	 a list of parameters for pmcorrect.method (if needed/wanted)

summary.method 	 the method used for the computation of expression values

summary.param 	 a list of parameters to be passed to the summary.method (if wanted).

summary.subset 	 a list of 'affyids'. If NULL, a expression summary value is computed for everything on the chip.

verbose 	 logical value. If TRUE it writes out some messages.

widget 	 a boolean to specify the use of widgets (the package tkWidget is required).

Details

Some arguments can be left to NULL if the widget=TRUE. In this case, a widget pops up and let the user choose with the mouse. The arguments are: AffyBatch, bgcorrect.method, normalize.method, pmcorrect.method and summary.method.

For the mas 5.0 and 4.0 methods ones need to normalize after obtaining expression. The function affy.scalevalue.exprSet does this.

To see what background methods are available you can look at the object bgcorrect.methods.

For the Li and Wong summary method notice you will not get the same results as you would get with dChip. dChip is not open source so it is not easy to reproduce. Notice also that this iterative algorithm will not always converge. If you run the algorithm on thousands of probes expect some non-convergence warnings. These are more likely when few arrays are used. We recommend using this method only if you have 10 or more arrays. Please refer to the fit.li.wong help page for more details.

Value

An object of class ExpressionSet, with an attribute pps.warnings as returned by the method computeExprSet.

See Also

AffyBatch

Examples


data(affybatch.example)

eset <- expresso(affybatch.example, bgcorrect.method="rma",
                 normalize.method="constant",pmcorrect.method="pmonly",
                 summary.method="avgdiff")

##to see options available for bg correction type:
bgcorrect.methods

[Package affy version 1.18.2 Index]