expressoWidget {affy}	R Documentation

A widget for users to pick correction methods

Description

This widget is called by expresso to allow users to select correction methods that will be used to process affy data.

Usage


expressoWidget(BGMethods, normMethods, PMMethods, expMethods, BGDefault,
normDefault, PMDefault, expDefault) 

Arguments

BGMethods 	 BGMethods a vector of character strings for the available methods that can be used as a background correction method of affy data

normMethods 	 normMethods a vector of character strings for the available methods that can be used as a normalization method of affy data

PMMethods 	 PMMethods a vector of character strings for the available methods that can be used as a PM correction method of affy data

expMethods 	 expMethodsa vector of character strings for the available methods that can be used as a summary method of affy data

BGDefault 	 BGDefault a character string for the name of a default background correction method

normDefault 	 normDefault a character string for the name of a default normalization method

PMDefault 	 PMDefault a character string for the name of a default PM correction method

expDefault 	 expDefault a character string for the name of a default summary method

Details

The widget will be invoked when expresso is called with argument "widget" set to TRUE. Default values can be changed using the drop down list boxes. Double clicking on an option from the dropdown list makes an selection. The first element of the list for available methods will be the default method if no default is provided.

Value

The widget returns a list of selected correction methods.

BG 	 background correction method

NORM 	 normalization method

PM 	 PM correction method

EXP 	 summary method

Author(s)

Jianhua Zhang

References

Documentations of affy package

See Also

expresso

Examples


if(interactive()){
  require(widgetTools)
  expressoWidget(c("mas", "none", "rma"), c("constant", "quantiles"),
c("mas", "pmonly"), c("liwong", "playerout"))
}

[Package affy version 1.18.2 Index]