hlog {affy}	R Documentation

Hybrid Log

Description

Given a constant c this function returns x if x is less than c and sign(x)*(c*log(abs(x)/c) + c) if its not. Notice this is a continuos odd ( f(-x)=-f(x) ) function with continous first derivative. The main purpose is to perform log transformation when one has negative numbers, for example for PM-MM.

Usage


hlog(x, constant=1)

Arguments

x 	 a number.

constant 	 the constant c (see description).

Details

If constant is less than or equal to 0 log(x) is returned for all x. If constant is infinity x is returned for all x.

Author(s)

Rafael A. Irizarry

[Package affy version 1.18.2 Index]