mas5 {affy}	R Documentation

MAS 5.0 expression measure

Description

This function converts an instance of AffyBatch into an instance of ExpressionSet using our implementation of Affymetrix's MAS 5.0 expression measure.

Usage


mas5(object, normalize = TRUE, sc = 500, analysis = "absolute", ...)

Arguments

object 	 an instance of AffyBatch

normalize 	 logical. If TRUE scale normalization is used after we obtain an instance of ExpressionSet

sc 	 Value at which all arrays will be scaled to.

analysis 	 should we do absolute or comparison analysis, although "comparison" is still not implemented.

... 	 other arguments to be passed to expresso.

Details

This function is a wrapper for expresso and affy.scalevalue.exprSet.

Value

ExpressionSet
The methods used by this function were implemented based upon available documentation. In particular a useful reference is Statistical Algorithms Description Document by Affymetrix. Our implementation is based on what is written in the documentation and as you might appreciate there are places where the documentation is less than clear. This function does not give exactly the same results. All source code of our implementation is available. You are free to read it and suggest fixes.
For more information visit this URL: http://stat-www.berkeley.edu/users/bolstad/

See Also

expresso,affy.scalevalue.exprSet

Examples


  data(affybatch.example)
  eset <- mas5(affybatch.example)

[Package affy version 1.18.2 Index]