mva.pairs {affy}	R Documentation

M vs. A Matrix

Description

A matrix of M vs. A plots is produced. Plots are made on the upper triangle and the IQR of the Ms are displayed in the lower triangle

Usage


mva.pairs(x, labels=colnames(x), log.it=TRUE,span=2/3,family.loess="gaussian",
          digits=3,line.col=2,main="MVA plot",cex=2,...)

Arguments

x 	 A matrix containing the chip data in the columns.

labels 	 the names of the variables.

log.it 	 logical. If TRUE uses log scale.

span 	 span to be used for loess fit.

family.loess 	 "gaussian" or "symmetric" as in loess.

digits 	 number of digits to use in the display of IQR.

line.col 	 color of the loess line.

main 	 an overall title for the plot.

cex 	 size for text

... 	 graphical parameters can be given as arguments to mva.plot

See Also

pairs

Examples


    
    x <- matrix(rnorm(4000),1000,4)
    x[,1] <- x[,1]^2
    dimnames(x) <- list(NULL,c("chip 1","chip 2","chip 3","chip 4"))
    mva.pairs(x,log=FALSE,main="example")

[Package affy version 1.18.2 Index]