normalize-methods {affy}	R Documentation

Normalize Affymetrix Probe Level Data - methods

Description

Method for normalizing Affymetrix Probe Level Data

Usage


normalize.methods(object)

Arguments

object 	 An AffyBatch.

Details

If object is an AffyBatch then normalize(object) returns an AffyBatch with the intensities normalized using the methodology specified by getOption("BioC")$affy$normalize.method. The affy package default is quantiles.

Other methodologies can be used by specifying them with the method argument. For example to use the invariant set methodology described by Li and Wong (2001) one would type: normalize(object, method="invariantset").

Further arguments passed by ..., apart from method, are passed along to the function responsible for the methodology defined by the method argument.

A character vector of nicknames for the methodologies available is returned by normalize.methods(object)), where object is an AffyBatch, or simply by typing normalize.AffyBatch.methods. If the nickname of a method is called "loess", the help page for that specific methodology can be accessed by typing ?normalize.loess.

For more on the normalization methodologies currently implemented please refer to the vignette `Custom Processing Methods'.

To add your own normalization procedures please refer to the customMethods vignette.

See Also

AffyBatch-class, normalize.

Examples


data(affybatch.example)
normalize.methods(affybatch.example)

[Package affy version 1.18.2 Index]