normalize.contrasts {affy}	R Documentation

Normalize intensities using the contrasts method

Description

Scale chip objects in an AffyBatch-class.

Usage


normalize.AffyBatch.contrasts(abatch,span=2/3,choose.subset=TRUE,
                              subset.size=5000,verbose=TRUE,
                              family="symmetric",type=c("together","pmonly","mmonly","separate"))

Arguments

abatch 	 an AffyBatch-class

span 	 parameter to be passed to the function loess.

choose.subset

subset.size

verbose 	 verbosity flag

family 	 parameter to be passed to the function loess.

type 	 A string specifying how the normalization should be applied.

Value

An object of the same class as the one passed.

See Also

maffy.normalize

[Package affy version 1.18.2 Index]