normalize.invariantset {affy}	R Documentation

Invariante Set normalization

Description

Normalize arrays in an AffyBatch using an invariant set.

Usage


normalize.AffyBatch.invariantset(abatch,
                   prd.td=c(0.003, 0.007), verbose=FALSE,baseline.type=c("mean","median","pseudo-mean","pseudo-median"),type=c("separate","pmonly","mmonly","together"))

normalize.invariantset(data, ref, prd.td=c(0.003,0.007))

Arguments

abatch 	 an AffyBatch

data 	 a vector of intensities on a chip (to normalize to the reference).

ref 	 a vector of reference intensities.

prd.td 	 cutoff parameter (details in the bibliographic reference)

baseline.type 	 Specify how to determine the baseline array

type 	 A string specifying how the normalization should be applied. See details for more.

verbose 	 A flag to have a dumps throughout the normalization

Details

The set of invariant intensities between data and ref is found through an iterative process (based on the respective ranks the intensities). This set of intensities is used to generate a normalization curve by smoothing.

The type argument should be one of "separate","pmonly","mmonly","together" which indicates whether to normalize only one probe type (PM,MM) or both together or separately.

Value

Respectively a AffyBatch of normalized objects, or a vector of normalized intensities, with an attribute "invariant.set" holding the indexes of the 'invariant' intensities.

Author(s)

L. Gautier <laurent@cbs.dtu.dk> (Thanks to Cheng Li for the discussions about the algorithm.)

References

Cheng Li and Wing Hung Wong, Model-based analysis of oligonucleotides arrays: model validation, design issues and standard error application. Genome Biology 2001, 2(8):research0032.1-0032.11

See Also

normalize to normalize AffyBatch objects.

[Package affy version 1.18.2 Index]