normalize.loess {affy}	R Documentation

Scale microarray data

Description

Normalizes arrays using loess.

Usage


normalize.loess(mat, subset = sample(1:(dim(mat)[1]), min(c(5000,
                 nrow(mat)))), epsilon = 10^-2, maxit = 1, log.it =
                 TRUE, verbose = TRUE, span = 2/3, family.loess =
                 "symmetric")
normalize.AffyBatch.loess(abatch,type=c("together","pmonly","mmonly","separate"), ...)

Arguments

mat 	 a matrix with columns containing the values of the chips to normalize.

abatch 	 an AffyBatch object.

subset 	 a subset of the data to fit a loess to.

epsilon 	 a tolerance value (supposed to be a small value - used as a stopping criterium).

maxit 	 maximum number of iterations.

log.it 	 logical. If TRUE it takes the log2 of mat

verbose 	 logical. If TRUE displays current pair of chip being worked on.

span 	 parameter to be passed the function loess

family.loess 	 parameter to be passed the function loess. "gaussian" or "symmetric" are acceptable values for this parameter.

type 	 A string specifying how the normalization should be applied. See details for more.

... 	 any of the options of normalize.loess you would like to modify (described above).

Details

The type arguement should be one of "separate","pmonly","mmonly","together" which indicates whether to normalize only one probe type (PM,MM) or both together or separately.

See Also

normalize

Examples


     #require(affy)
     #data(Dilution)
     #x <- pm(Dilution[,1:3])
     #mva.pairs(x)
     #x <- normalize.loess(x,subset=1:nrow(x))
     #mva.pairs(x)

[Package affy version 1.18.2 Index]