pairs.AffyBatch {affy}	R Documentation

plot intensities using 'pairs'

Description

Plot intensities using the function 'pairs'

Usage


## S3 method for class 'AffyBatch':
pairs(x, panel=points, ..., transfo=I, main=NULL, oma=NULL,
                  font.main = par("font.main"),
                  cex.main = par("cex.main"), cex.labels = NULL, 
                  lower.panel=panel, upper.panel=NULL, diag.panel=NULL,
                  font.labels = 1, row1attop = TRUE, gap = 1)

Arguments

x 	 an AffyBatch object

panel 	 a function to produce a plot (see pairs)

... 	 extra parameters for the 'panel' function

transfo 	 a function to transform the intensity values before generating the plot. 'log' and 'log2' are popular choices.

main 	 title for the plot

oma 	 see 'oma' in par.

font.main 	 see pairs

cex.main 	 see pairs

cex.labels 	 see pairs

lower.panel 	 a function to produce the plots in the lower triangle (see pairs).

upper.panel 	 a function to produce the plots in the upper triangle (see pairs).

diag.panel 	 a function to produce the plots in the diagonal (see pairs).

font.labels 	 see pairs

row1attop 	 see pairs

gap 	 see pairs

Details

Plots with several chips can represent zillions of points. They require a lot of memory and can be very slow to be displayed. You may want to try to split of the plots, or to plot them in a device like 'png' or 'jpeg'.

[Package affy version 1.18.2 Index]