plotLocation {affy}	R Documentation

Plot a location on a cel image

Description

Plots a location on a previously plotted cel image. This can be used to locate the physical location of probes on the array.

Usage


plotLocation(x, col="green", pch=22, ...)

Arguments

x 	 a `location'. It can be obtained by the method of AffyBatch indexProbes, or made elsewhere (basically a location is nrows and two columns array. The first column corresponds to the x positions and the second columns corresponds to the y positions of n elements to locate)

col 	 colors for the plot

pch 	 plotting type (see function plot)

... 	 Other parameters passed to the function points

Author(s)

Laurent

See Also

AffyBatch

Examples


  ## loading data
  data(affybatch.example)

  ## image of the celfile
  image(affybatch.example[, 1])

  ## genenames, arbitrarily pick the 101th
  n <- geneNames(affybatch.example)[101]

  ## get the location for the gene n
  l <- indexProbes(affybatch.example, "both", n)[[1]]
  ## convert the index to X/Y coordinates
  xy <- indices2xy(l, abatch=affybatch.example) 

  ## plot
  plotLocation(xy)


[Package affy version 1.18.2 Index]