pmcorrect {affy}	R Documentation

PM Correction

Description

Corrects the PM intensities in a ProbeSet for nonspecific binding.

Usage


pmcorrect.pmonly(object)

pmcorrect.subtractmm(object)

pmcorrect.mas(object, contrast.tau=0.03, scale.tau=10, delta=2^(-20))

Arguments

object 	 An object of class ProbeSet.

contrast.tau 	 a number denoting the contrast tau parameter in the MAS 5.0 pm correction algorithm.

scale.tau 	 a number denoting the scale tau parameter in the MAS 5.0 pm correction algorithm.

delta 	 a number denoting the detla parameter in the MAS 5.0 pm correction algorithm.

Details

These are the pm correction methods perfromed by Affymetrix MAS 4.0 (subtractmm) and MAS 5.0 (mas). See the Affymetrix Manual for details. pmonly does what you think: does not change the PM values.

Value

A ProbeSet for which the pm slot contains the corrected PM values.

References

Affymetrix MAS 4.0 and 5.0 manual

Examples


data(affybatch.example)
gn <- geneNames(affybatch.example)
pps <- probeset(affybatch.example, gn[1])[[1]]

pps.pmonly <- pmcorrect.pmonly(pps)
pps.subtractmm <- pmcorrect.subtractmm(pps)
pps.mas5 <- pmcorrect.mas(pps)
par(mfrow=c(2,2))
#plot(pm(pps.pmonly), pm(pps.subtractmm))
#plot(pm(pps.pmonly),pm(pps.mas5))
#plot(pm(pps.subtractmm),pm(pps.mas5))

[Package affy version 1.18.2 Index]