read.affybatch {affy}	R Documentation

Read CEL files into an AffyBatch

Description

Read CEL files into an Affybatch.

Usage


read.affybatch(..., filenames = character(0),
               phenoData = new("AnnotatedDataFrame"),
               description = NULL,
               notes = "",
               compress = getOption("BioC")$affy$compress.cel,
               rm.mask = FALSE, rm.outliers = FALSE, rm.extra = FALSE,
               verbose = FALSE,sd=FALSE, cdfname = NULL)

ReadAffy(..., filenames=character(0),
              widget=getOption("BioC")$affy$use.widgets,
              compress=getOption("BioC")$affy$compress.cel,
              celfile.path=NULL,
              sampleNames=NULL,
              phenoData=NULL,
              description=NULL,
              notes="",
              rm.mask=FALSE, rm.outliers=FALSE, rm.extra=FALSE,
              verbose=FALSE,sd=FALSE, cdfname = NULL)

Arguments

... 	 file names separated by comma.

filenames 	 file names in a character vector.

phenoData 	 a AnnotatedDataFrame object.

description 	 a MIAME object.

notes 	 notes.

compress 	 are the CEL files compressed?

rm.mask 	 should the spots marked as 'MASKS' set to NA?

rm.outliers 	 should the spots marked as 'OUTLIERS' set to NA?

rm.extra 	 if TRUE, then overrides what is in rm.mask and rm.oultiers.

verbose 	 verbosity flag.

widget 	 a logical specifying if widgets should be used.

celfile.path 	 a character denoting the path ReadAffy should look for cel files.

sampleNames 	 a character vector of sample names to be used in the AffyBatch.

sd 	 should the standard deviation values in the CEL file be read in? Since these are typically not used default is not to read them in. This also save lots of memory.

cdfname 	 used to specify the name of an alternative cdf package. If set to NULL, then the usual cdf package based on Affymetrix' mappings will be used.

Details

ReadAffy is a wrapper for read.affybatch that permits the user to read in phenoData, MIAME information, and CEL files using widgets. One can also define files where to read phenoData and MIAME information.

If the function is called with no arguments ReadAffy() all the CEL files in the working directory are read and put into an AffyBatch. However, the arguments give the user great flexibility.

phenoData is read using read.AnnotatedDataFrame. If a character is given it tries to read the file with that name to obtain the AnnotationDataFrame object as described in read.AnnotatedDataFrame. If left NULL and widget=FALSE (widget=TRUE is not currently supported), then a default object is created. It will be an object of class AnnotatedDataFrame with its pData being a data.frame with column x indexing the CEL files.

AllButCelsForReadAffy is an internal function that gets called by ReadAffy. It gets all the information except the cel intensities.

description is read using read.MIAME. If a character is given, then it tries to read the file with that name to obtain a MIAME instance. If left NULL but widget=TRUE, then widgets are used. If left NULL and widget=FALSE, then an empty instance of MIAME is created.

Value

An AffyBatch object.

Author(s)

Ben Bolstad bmb@bmbolstad.com (read.affybatch), Laurent Gautier, and Rafael A. Irizarry (ReadAffy)

See Also

AffyBatch

Examples


if(require(affydata)){
     celpath <- system.file("celfiles", package="affydata")
     fns <- list.celfiles(path=celpath,full.names=TRUE)
 
     cat("Reading files:\n",paste(fns,collapse="\n"),"\n")
     ##read a binary celfile
     abatch <- ReadAffy(filenames=fns[1])
     ##read a text celfile
     abatch <- ReadAffy(filenames=fns[2])
     ##read all files in that dir
     abatch <- ReadAffy(celfile.path=celpath)
}

[Package affy version 1.18.2 Index]