xy2indices {affy}	R Documentation

Functions to convert indices to x/y (and reverse)

Description

Functions to convert indices to x/y (and reverse)

Usage


xy2indices(x, y, nr = NULL, cel = NULL, abatch = NULL, cdf = NULL, xy.offset = NULL)
indices2xy(i, nr = NULL, cel = NULL, abatch = NULL, cdf = NULL, xy.offset = NULL)

Arguments

x 	 X position for the probes

y 	 Y position for the probes

i 	 indices in the AffyBatch for the probes

nr 	 total number of Xs on the chip

cel 	 a corresponding object of class Cel

abatch 	 a corresponding object of class AffyBatch

cdf 	 character - the name of the corresponding cdf package

xy.offset 	 an eventual offset for the XY coordinates. See Details

Details

The probes intensities for given probe set ids are extracted from an AffyBatch object using the indices stored in the corresponding cdfenv.

The parameter xy.offset is there for compatibility. For historical reasons, the xy-coordinates for the features on Affymetrix chips were decided to start at 1 (one) rather than 0 (zero). One can set the offset to 1 or to 0. Unless the you _really_ know what you are doing, it is advisable to let it at the default value NULL. This way the package-wide option code{xy.offset} is always used.

Value

A vector of indices or a two-columns matrix of Xs and Ys.

Warning

Even if one really knows what is going on, playing with the parameter xy.offset could be risky. Changing the package-wide option xy.offset appears much more sane.

Author(s)

L.

See Also

indexProbes

Examples


data(affybatch.example)

pm.i <- indexProbes(affybatch.example, which="pm", genenames="AFFX-BioC-5_at")[[1]]
mm.i <- indexProbes(affybatch.example, which="mm", genenames="AFFX-BioC-5_at")[[1]]

pm.i.xy <- indices2xy(pm.i, abatch = affybatch.example)
mm.i.xy <- indices2xy(mm.i, abatch = affybatch.example)

image(affybatch.example[1], transfo=log2)
## plot the pm in red
plotLocation(pm.i.xy, col="red")
plotLocation(mm.i.xy, col="blue")


[Package affy version 1.18.2 Index]